test(racingcar): Referee 테스트 분리
